Update rocket to latest rc and fix changes. Reintroduce Cargo.lock
Since there was no Cargo.lock, cargo was getting rocket 0.5.0-rc.3 instead of 0.5.0-rc.1,
and this broke the build (I also tried forcing the version that then conflicted with some other rocket's internal
crates so I gave up). Cargo.lock needs to be part of the repo so that everyone
gets the same versions of dependencies (that are known to work).

Then I had to do some changes to make the examples work with 0.5.0-rc.3. I tried
making them as similar as possible to what is shown at the the current documentation for rocket.

I was getting a weird error saying I wasn't allowed to use `extern crate` in the example from lib.rs
that I wasn't able to solve, so I instead replaced it with a `use`