Use tree_magic_mini instead of tree_magic (to fix future incompatibility warning)
When building this crate, cargo issues a warning saying there's a future incompatibility in one
of the dependencies (an old version of the nom crate which is used by tree_magic).

Fixing it requires that tree_magic updates the version of nom that is uses. Since it is
no longer maintained (the github repo got recently archived) I decided to switch to a fork
(tree_magic_mini) that fixes these issues. It is also more widely used:
https://lib.rs/crates/tree_magic
https://lib.rs/crates/tree_magic_mini